Added search:indices rake task for introspection
Namespacing is not properly handled as the `search:indices` task is
mounted on `app:search:indices` within the engine project, but once
inside an application that mounts this engine, the rake task appears as
`search:indices` instead. Not a minor issue.

Had some trouble with double execution of the rake task which most likely
has to do with the manner in which the bootstrapping is handled. I have
a `/lib/gemname/railtie.rb` file which still needs to be commited on my
current setup. I noticed that altering the settings in the
/lib/gemname/railtie.rb` and `gemname.gemspec` altered this behavior but
I'm racing against time and as such I deeem this more of a cosmetic
blemish.